search added and some improved

+++ frontend/app/controllers/books.js
@@ -7,10 +7,9 @@ export default Controller.extend({
     data:EmberObject.create(),
 
     init(){
-      console.log("init");
-      this.set('data', this.send('sorgula', ''));
+      this.set('data', this.send('queryByTitle', ''));
     },
-    
+
     actions: {
       addBook() {
         let self = this;
@@ -23,30 +22,29 @@ export default Controller.extend({
         self.transitionToRoute("/update/" + id);
       },
 
-      delete(id) {
-        let self = this;
-        Ember.$.ajax({
-            url: `http://localhost:8080/delete/${id}`,
-            type: "DELETE",
-            headers: {
-                'Content-Type': 'application/json'
-              }
-          }).then(function(resp){
-            console.log("successful");
-            window.location.reload(true);
-          }).catch(function(error){
-            console.log(error);
-            console.log("negative");
-          });
+      delete(book) {
+        if(confirm('You are about to delete "' + book.title + "\"\n\nAre you sure?")){
+          let self = this;
+          Ember.$.ajax({
+              url: `http://localhost:8080/delete/${book.id}`,
+              type: "DELETE",
+              headers: {
+                  'Content-Type': 'application/json'
+                }
+            }).then(function(resp){
+              window.location.reload(true);
+            }).catch(function(error){
+            });
+        }
+
 
       },
-      sorgula(kriter){
-        console.log(kriter);
+      queryByTitle(query){
         var myURL = ""
-        if(kriter === "")
+        if(query === "")
           myURL = 'http://localhost:8080/books';
         else {
-          myURL = 'http://localhost:8080/books/title/' + kriter;
+          myURL = 'http://localhost:8080/books/title/' + query;
         }
         Ember.$.ajax({
           url:myURL,
